snagflash: fastboot: generalise handling of command arguments
Use the str.split() function to split the command from its arguments,
since that doesn't require us to check for the ':' character.

The object returned from str.split() is as list, so partition that into
the command, and a list of its optional arguments. Then use the fact
that the arguments is a list, to pass it to the fastboot method as a
varying argument.

Signed-off-by: Martin Hundebøll <martin@geanix.com>