src: snagrecover: imx: Wrap HID device in SDP object before calling run_firmware
As of now, the firmware handling layer receives an HID device and wraps it in an
SDPCommand object. However, direct access to the HID device is not needed in the
firmware handling procedures, and instanciating USB protocol devices should be
the responsibility of the recoveries layer, not the firmware layer.

Additionally, add a close() method to the SDPCommand class so that the
recoveries layer can use the SDPCommand instance instead of directly accessing
the underlying HID device.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>