snagfactory: core module for new factory flashing tool
Snagfactory is a new tool that wraps around snagrecover and snagflash to
provide parallel recovery and flashing of multiple boards.

The end goal is to have an efficient factory flashing tool complete with a
GUI that can be used on both Linux and Windows.

This first module defines basic methods and data structures to manage
sequential recovery and flashing stages in a way that can be parallelized
for multiple boards.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>