snagfactory: add Kivy GUI
Add a GUI for snagfactory.

The core of this interface is a list of board widgets that allow an
operator to monitor the flashing status of each board defined in the
factory batch.

Snagfactory will regularly scan for boards in recovery mode using the provided
vid:pid pairs. Once the start order is given, the sequential recovery and
flashing tasks for each board will be run in parallel.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>