snagrecover: protocols: fastboot: allow downloading a specific section of a file
Add an optional set of arguments to the Fastboot download command that
allows the caller to specify an offset and size in addition to the
mandatory path of the file to download.

This will be useful for snagfactory, to download huge files in multiple
sections without creating a new file for each section.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>