snagfactory: add partition table and huge image handling
The current format for batch file flashing sections allows specifying
images to be flashed to existing partitions. Change this to allow
specifying either one image to be flashed to the entire MMC, or a
partition table where each partition can have an associated image.

Add support for flashing images that are larger than the capacity of the
U-Boot Fastboot buffer. This is done by dividing the huge image into
multiple sections.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>