docs: describe eMMC hardware partitioning in U-Boot
eMMC hardware partitioning is a complex but potentially useful feature that
deserves some detailed documentation. It can currently be partially
implemented with Snagfactory using post-flashing commands, but a more
integrated support would be nice to have.

Document the eMMC hardware partitioning process in U-Boot and draft a
potential support for this feature in Snagfactory.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>