snagfactory: drop 'batch' terminology for configuration file
This terminology was initially chosen due to boards being processed in
separate "batches". However, given the direction that Snagfactory is
evolving in, this is no longer going to be relevant.

Drop "batch configuration file" terminology and simply use the term
"configuration file".

Additionally, fix some line endings in session.py and config.kv.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>