snagflash: add interactive mode
The flashing logic in Snagfactory is becoming quite complicated, and it
makes sense at this point to move most of it to Snagflash.

Introduce an interactive mode that defines a set of flashing tasks
that can be leveraged for Snagfactory.

Additionally, stop relying on the Fastboot flashing backend for writing
downloaded files to the storage device, and use "oem_run" instead. This
will grant Snagflash more flexibility and emancipate it from the
supported Fastboot backends. It will also allow Snagflash to handle
image sparsity on its own, without interference from U-Boot.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>