snagfactory: add rescan button
Currently, the only way to restart a scanning phase after finishing a
group of boards is to reaload the configuration file, which isn't very
practical.

Modify the "start" button so that pressing it after a group of boards
are done starts a new scanning phase with the same configuration file.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>