snagrecover: protocols: hid: drop HID-specific descriptor parsers
The current HID protocol implementation parses HID and Report
descriptors obtained from the target device. This is only used to determine
report direction when calling set_report() or get_report(). On Windows, it
is not possible to obtain raw HID device descriptors through the libusb
API, which breaks this parsing feature.

Remove this parsing logic from the hid module and hardcode report
directions as "output", since input reports are not used by Snagboot. This
allows the hid module to function properly on Windows and simplifies the
code drastically.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>