treewide: make snagfactory an optional package feature
The Kivy dependency is causing issues for some users. These users are not
necessarily interested in snagfactory and only want to use snagboot and
snagflash.

Add an optional "gui" dependency group and move the Kivy dependency to this new
group. Snagfactory will not be available unless this dependency group is
selected during installation.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>