snagflash: add separate protocol for extended Fastboot commands
The "fastboot" procotol of Snagflash currently provides basic
bootloader-agnostic Fastboot commands with the "-f" flag, as well as
extended U-Boot-specific commands in interactive mode with the "-i" and
"-I" flags.

This is unecessarily confusing, as it doesn't explicitely indicate that
"interactive mode commands" are U-Boot-specific.

To clarify this, move the "interactive mode commands" to a new protocol
named "fastboot-uboot", and deprecate usage of the "-i" and "-I" flags with
the "fastboot" protocol.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>