snagrecover: sunxi: bump pylibfdt version to 1.7.2.post1
This version bump is required to fix an issue caused by incompatibilities
between Swig 4.3 and previous versions of Pylibfdt. This issue affects all
Snagboot installations where Swig 4.3 was available on Pypi during the build.
It results in snagrecover failing on Allwinner Sunxi platforms because of an
exception raised in libfdt.getprop().

Additionally, the swig build-time dependency is now explicitely given in
Pylibfdt's pyproject.toml, so specifying it in Snagboot isn't necessary
anymore.

Signed-off-by: Romain Gantois <romain.gantois@bootlin.com>