snagrecover: firmware: make firmware format check optional in load_fw()
The load_fw() function currently consistently checks the format of the
requested firmware file, and emits a warning if it looks like a text file. Or,
the pending snagrecover support for BCM SoCs requires loading a firmware file
which is actually expected to be in text format. Thus, a non-relevant warning
will be emitted in this case.

Allow making this firmware file format check optional to avoid non-relevant
warnings when a text format is expected.

Signed-off-by: François Foltete <francois.foltete@bootlin.com>