README.md: improve the presentation of the list of SoC supported
So far, the list of SoC support was not very well presented, almost a
bit hidden, while it is a major feature of Snagboot that it supports
many different platforms.

Make this a little bit more prominent by using a bullet point list,
adding links, listing more exhaustively the supported platforms.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>